+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ### Changes
-- Check if the Stonecutter is added to the project before injecting code fragments.
+- Made Kotlin plugin dependency optional.